Fixed bug and made code better Fixed intro still need episode 2-5

+++ Match3LoveSim.spritebuilder/Source/MainScene.h
@@ -1,5 +1,5 @@
 @interface MainScene : CCNode
 
-@property (nonatomic, assign) int dayCounter;
+@property (nonatomic, assign) long dayCounter;
 
 @end
